+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [#1917](https://github.com/influxdb/influxdb/pull/1902): Creating Infinite Retention Policy Failed.
 - [#1758](https://github.com/influxdb/influxdb/pull/1758): Add Graphite Integration Test.
 - [#1929](https://github.com/influxdb/influxdb/pull/1929): Default Retention Policy incorrectly auto created.
+- [#1930](https://github.com/influxdb/influxdb/pull/1930): Auto create database for graphite if not specified.
 
 ### Features
 - [#1902](https://github.com/influxdb/influxdb/pull/1902): Enforce retention policies to have a minimum duration.